fix: add ::integer casts to COUNT(*) aggregates in service SQL
PostgreSQL COUNT(*) returns bigint serialized as string by the pg driver.
Without explicit ::integer casts, comparisons and arithmetic produce
unexpected results (e.g. '2' + '1' = '21' instead of 3).

Fixes: insights-store (dedup cnt), webhook-service (delivery count),
incident-store (active+resolved total addition).

+++ backend/src/services/insights-store.ts
@@ -120,7 +120,7 @@ export async function insertInsights(insights: InsightInsert[]): Promise<Set<str
   `;
 
   const dedupeSQL = `
-    SELECT COUNT(*) as cnt FROM insights
+    SELECT COUNT(*)::integer as cnt FROM insights
     WHERE container_id = ? AND category = ? AND title = ?
       AND created_at >= NOW() - INTERVAL '60 minutes'
   `;
